refactor(frontend): split active room cache
